Fix rotation of tower towards target, movement of projectile to target with support for static and moving targets
Fix rotation of tower so that it correctly points towards a target's position when it launches a projectile at that target.

Fix movement of projectile to its target. Projectile will now travel to static and moving targets by adjusting its flight path to the target as necessary while the target exists. If the target ceases to exist before the projectile reaches it, the projectile is removed from the level.

Remove impact of friction and gravity on projectiles.

Add extra rocket tower for moving target attack demonstration purposes. Adjust tower placements and enemy target speed for demonstration purposes.