Add limit on attack range of tower
Add limit on attack range of tower. The tower's range of attack is limited by its "range" attribute. The tower will check for all targets existing in the game world, and among those targets that are within range, it will attack the target that is closest to the tower.

Remove "RocketTower" and "GreenTower" entities as they are no longer needed for POC demonstration.