Change bomb explosion effects to improve performance on low-end hardware
Change bomb explosion effects to an effect similar to that used for stun tower attacks to improve performance on low-end hardware.
The flashier bomb explosion effects code is commented out for now. That code will need to be revisited to improve its efficiency, or removed if there is not enough bandwidth for more work on it.

Change bomb damage distribution within explosion radius to use the bomb's target's position (blastSite) instead of the bomb projectile position when calculating the distance of enemies from the point of impact to determine if they are within the explosion radius.