Add draft HUD life indicator to bottom right corner of play screeen
Add a draft version of the HUD life indicator to the bottom right corner of the play screen that shows how many lives the player has. Once the player's life value reaches zero or lower, the HUD displays the text "Game Over" below the life indicator.

The melonJS platformer tutorial's "PressStart2P" font was used for the HUD font.