Extend game over screen to support allowing the player to restart the current level if they are past level01
Extend the game over screen to support allowing the player choose to restart the current level if they are past level01, or to restart the entire game by going to the title screen. The player's gold and life balance are saved at the end of each level to enable this feature.