Update tower and related projectiles code to minimize use of "getChildBy" calls
Update tower and related projectiles code to minimize use of "getChildBy" calls when searching for target entities and position marker objects.

Enemy targets are now searched for using game.data.enemies, and static position marker objects are now tracked using a local variable in the projectile that is set during initialization of the projectile or when the static position marker object is created during the projectile's update function.