Update game over and game end screens to play and stop music using play() with loop option set and stop()
Update game over and game end screens to play and stop looping music using play() with loop option set and stop() instead of the playTrack() and stopTrack() functions, the latter of which do not work consistently on the live site.

This change works locally (the music loops at those screens as expected) on Ricardo's computer, however it is not known if this change will ensure music looping works at those screens as expected on the live site and testing is needed to check that.