Update pause menu icon button
Update the pause menu icon button such that when the pause menu is open, the icon displayed is a play button (instead of a pause button). This change is intended to make it more clear to the user that if they click the pause menu icon button again, the pause menu will close and the game will resume.